Added ABC for streaming and class for sim'ed (#26)
In the new file SimulatedTimeSeries.py, there are two new classes.

There is an abstract base class StreamTimeSeriesInterface.
This defines only the produce method.

There is SimulatedTimeSeries, which takes in a generator of time
series data and can produce chunks of the time series.

Both of these are documented in issue #26.