Updated FileStorageManager functionality
It is now possible to open a new FileStorageManager and load the index of the existing manager in that directory.  This gives the new storage manager access to the existing index.

However, the two managers are now competing for the same index file and are overwriting each other's changes without communicating. This is a problem.  This change is meant to allow a new manager to read the files from another manager then exit without modifying anything.